Move test_echo_server into tcp test file
